requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
DAWG-Python==0.7.2
docopt==0.6.2
fake-useragent==2.2.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.5.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
mysql-connector-python==8.3.0
networkx==3.5
numpy==1.26.0
packaging==25.0
pandas==2.1.1
pillow==10.3.0
pip==25.1.1
propcache==0.3.1
protobuf==4.21.12
pydantic==2.11.5
pydantic_core==2.33.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.5
pytz==2025.2
PyYAML==6.0.2
pyzbar==0.1.9
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==70.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
yarl==1.20.0
